command-palette: simplify example for iterable commands
The example for iterable commands combines an iterable command with a
nested command. Without reading the section about nested commands, it's
not immediately obvious that these are separate concepts.

Change the example for the iterable command so that it does not also use
a nested command.

Add a second, separate example that shows combining nested and iterable
commands.