json-fragment-extensions: output GUIDs in "registry format"
Windows Terminal expects the 'guid' property of a profile in the
so-called "registry format" [1], i.e. enclosed in curly brackets.

The different GUID examples in TerminalDocs/json-fragment-extensions.md
are enclosed in curly brackets, but the Python script example to
generate GUIDs prints them without the enclosing curly brackets.

Use a Python f-string to print them enclosed in curly brackets.

[1] https://docs.microsoft.com/en-us/windows/terminal/customize-settings/profile-advanced#unique-identifier