troubleshooting: remove confusing 'Important' note about WSL 'startingDirectory'
In e7f739e (Docs updates for 1.11 release (#397), 2021-08-31), an
"Important" note was added to the "Set your WSL distribution to start in
the home directory" section of the troubleshooting page to mention that
'startingDirectory' now accepts Linux-style paths, following
microsoft/terminal@a2a6050 (When launching wsl, promote the starting
directory to --cd (#9223), 2021-08-02).

The note mentions that Linux-styled paths were accepted "on newer
versions of Windows", presumably meaning newer than 1903. Indeed, that
version was at the time mentioned in the text as the first Windows
vesion supporting the '//wsl$/' P9 prefix (it had been mentioned since
the WSL startingDirectory troubleshooting entry was created in da815cc
(Review troubleshooting page, 2020-05-04).

The text of the section was updated in 7cd48bc (startingDirectory path
for WSL no longer requires prefix, 2022-01-07) to add an example with a
Linux-style path, and to mention that this feature was added in Windows
Terminal itself (not in Windows), thus removing the mention of Windows
1903. However, the "Important" note was left, and it still confusingly
mentions that Linux-style paths are accepted on "newer" versions of
Windows, without mentioning newer than what.

Remove the note altogether.