troubleshooting: Linux paths support in 'startingDirectory' was added in 1.11.2921.0
In 7cd48bc (startingDirectory path for WSL no longer requires prefix,
2022-01-07), the "Set your WSL distribution to start in the home
directory" section of the troubleshooting guide was updated to mention
that WSL paths can now be directly referenced, following
microsoft/terminal@a2a6050 (When launching wsl, promote the starting
directory to --cd (#9223), 2021-08-02).

The text mentions this feature was added in Windows Terminal
1.12.3472.0, but it actually dates back to v1.11.2421.0 (Preview) and
v1.11.2921.0 (Stable).

Update the text to mention the stable version, and link to this version
in the GitHub releases. Also slightly improve the wording (s/are now/can
now be/).